+++ client/lib/pages/game/resources/player_position.dart
@@ -3,7 +3,7 @@ import 'package:latlong2/latlong.dart';
 
 class PlayerState {
   final LatLng position;
-  final int health;
+  final double health;
 
   PlayerState({
     required this.position,
