+++ server/pkg/zone/zone.go
@@ -15,12 +15,12 @@ const (
 type Zone struct {
 	Position s2.LatLng
 	// Zone radius in meters
-	Radius uint32
+	Radius float32
 	// Zone damage per period
 	Damage float32
 }
 
-func NewZone(position s2.LatLng, radius uint32, minZoneRadiusInMeter uint32) *Zone {
+func NewZone(position s2.LatLng, radius float32, minZoneRadiusInMeter float32) *Zone {
 	z := new(Zone)
 
 	z.Position = position
@@ -31,7 +31,7 @@ func NewZone(position s2.LatLng, radius uint32, minZoneRadiusInMeter uint32) *Zo
 }
 
 // Calc damage to have target damage at last zone
-func calcDamage(radius uint32, minZoneRadiusInMeter uint32) float32 {
+func calcDamage(radius float32, minZoneRadiusInMeter float32) float32 {
 	if radius > minZoneRadiusInMeter {
 		zonesCount := math.Log2(float64(radius)) - math.Log2(float64(minZoneRadiusInMeter))
